+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - chore(packages): Update packages and fix breaking changes
+- refactor(core): Add `bulkSize` option to `bulkIndex` method
 
 ## [1.3.2] 2019-01-30